Install perf dependencies on setup
