+++ crates/infra/cli/src/commands/perf/mod.rs
@@ -1,4 +1,5 @@
 mod archive;
+mod binaries;
 mod cargo;
 mod npm;
